Move activateDecorations to own file.
